requirements.txt
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-money==3.5.3
django-phonenumber-field==8.0.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.48
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
py-moneyed==3.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
